requirements.txt
# Async & Infrastructure
redis>=4.5.0

# Stream Processing
opencv-python>=4.8.0
av>=10.0.0

# ML/AI Libraries
librosa>=0.10.0
openai-whisper>=20230314
transformers>=4.30.0
mediapipe>=0.10.0,<0.10.8  # Versions <0.10.8 work with Python 3.9
numpy>=1.24.0,<2.0.0  # Pin to 1.x for compatibility
torch>=2.0.0

# UI & Testing
streamlit>=1.25.0
pytest>=7.4.0
hypothesis>=6.82.0

# Additional utilities
pyyaml>=6.0
